demo-drf/pyproject.toml
[tool.poetry]
name = "demo-drf"
version = "0.1.0"
description = ""
authors = ["Arpan Koirala <arpan.koirala@yco.com.np>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
